requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.0
blis==0.7.11
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Django==5.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.4.0
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
preshed==3.0.9
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
sympy==1.13.2
tbb==2021.13.1
thinc==8.2.5
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
